+++ src/tapmap/ui/modal_view.py
@@ -573,7 +573,10 @@ class ModalTextBuilder:
                                         ],
                                     ),
                                     html.P(
-                                        "Stored securely in your computer's keyring."
+                                            "Stored in the Docker data directory for future updates."
+                                            if is_docker
+                                            else "Stored securely in your computer's keyring "
+                                                 "for future updates."
                                     ),
                                     html.Button(
                                         "Install MaxMind database files",